requirements.txt
asgiref==3.5.0
autopep8==1.5.6
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
pipreqs==0.4.11
pycodestyle==2.7.0
PyJWT==1.7.1
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
wincertstore==0.2
yarg==0.1.9
PyMyMQL==1.0.2
mysqlclient==2.1.0
